+++ tools/xstore/cli/restore.py
@@ -32,6 +32,8 @@ from core.backup_restore.storage.filestream_client import FileStreamClient, Back
 from core.backup_restore.utils import check_run_process
 import wget
 import requests
+from .common import check_parameters_exist, get_parameter_value
+
 
 RESTORE_TEMP_DIR = "/data/mysql/restore"
 CONN_TIMEOUT = 30
@@ -57,8 +59,13 @@ def start(restore_context):
         sink = params["sink"]
         pitr_endpoint = params["pitrEndpoint"] if "pitrEndpoint" in params else ""
         pitr_xstore = params["pitrXStore"] if "pitrXStore" in params else ""
+        is_pxc_xstore = params["pxcXStore"]
+        keyring_path = params["keyringPath"] if "keyringPath" in params else ""
+        keyringfile_path = params["keyringFilePath"] if "keyringFilePath" in params else ""
 
-    logger.info('start restore: backup_file_path=%s' % backup_file_path)
+    logger.info('start restore: commit_index=%s, backup_file_path=%s, isPXCXStore:%s, pitr_endpoint=%s,'
+                'pitr_xstore=%s,keyring_path=%s'
+                % (commit_index, backup_file_path, is_pxc_xstore, pitr_endpoint, pitr_xstore,keyring_path))
 
     context = Context()
     node_role = context.node_role()
@@ -67,6 +74,8 @@ def start(restore_context):
         return
 
     filestream_client = FileStreamClient(context, BackupStorage[str.upper(storage_name)], sink)
+
+    keyring_path_local = download_keyring_file(keyringfile_path,keyring_path, filestream_client, logger)
 
     mkdir_needed(context)
 
@@ -78,38 +87,72 @@ def start(restore_context):
 
     initialize_local_mycnf(context, logger)
 
-    apply_backup_file(context, logger)
+    apply_backup_file(keyring_path_local, context, logger)
 
-    mysql_bin_list = download_binlogbackup_file(binlog_dir_path, filestream_client, logger) if len(
-        pitr_endpoint) == 0 else download_pitr_binloglist(context, pitr_endpoint, pitr_xstore, logger)
+    if is_pxc_xstore:
+        mysql_bin_list = download_binlogbackup_file(binlog_dir_path, filestream_client, logger) if len(
+            pitr_endpoint) == 0 else download_pitr_binloglist(context, pitr_endpoint, pitr_xstore, logger)
 
-    copy_binlog_to_new_path(mysql_bin_list, context, logger)
+        copy_binlog_to_new_path(mysql_bin_list, context, logger)
 
-    cluster_start_index = get_xtrabackup_binlog_info_from_instance_local(context)
-    logger.info("cluster_start_index is: %s" % cluster_start_index)
+        cluster_start_index = get_xtrabackup_binlog_info_from_instance_local(context)
+        logger.info("cluster_start_index is: %s" % cluster_start_index)
 
-    chown_data_dir(context, logger)
+        chown_data_dir(context, logger)
 
-    last_binlog, first_binlog = show_last_and_first_binlog(context, logger)
+        last_binlog, first_binlog = show_last_and_first_binlog(context, logger)
 
-    end_index, end_term = xdb_show_binlog_index(last_binlog, context, logger)
-    logger.info("end_index:%s;end_term:%s" % (end_index, end_term))
+        end_index, end_term = xdb_show_binlog_index(last_binlog, context, logger)
+        logger.info("end_index:%s;end_term:%s" % (end_index, end_term))
 
-    init_mysqld_metadata(cluster_start_index, commit_index, context, end_term, node_role, logger)
+        init_mysqld_metadata(cluster_start_index, commit_index, context, end_term, node_role, logger, is_pxc_xstore,
+                             pitr_endpoint)
 
-    p = subprocess.Popen([
-        os.path.join(context.engine_home, 'bin', 'mysqld'),
-        "--defaults-file=" + context.mycnf_path,
-        "--user=mysql"
-    ], stdout=sys.stdout)
+        p = subprocess.Popen([
+            os.path.join(context.engine_home, 'bin', 'mysqld'),
+            "--defaults-file=" + context.mycnf_path,
+            "--user=mysql"
+        ], stdout=sys.stdout)
 
-    wait_binlog_apply_ready(context.port_access(), end_index, logger)
+        wait_binlog_apply_ready(context.port_access(), end_index, logger)
 
-    p.kill()
-    p.wait()
+        p.kill()
+        p.wait()
 
+    else:
+        if len(pitr_endpoint) != 0:
+            mysql_bin_list = download_pitr_binloglist(context, pitr_endpoint, pitr_xstore, logger)
+            copy_binlog_to_new_path(mysql_bin_list, context, logger)
+            cluster_start_index = get_xtrabackup_binlog_info_from_instance_local(context)
+            logger.info("cluster_start_index is: %s" % cluster_start_index)
+
+            chown_data_dir(context, logger)
+
+            last_binlog, first_binlog = show_last_and_first_binlog(context, logger)
+            pitr_ts = download_pitr_ts(pitr_endpoint, logger)
+            truncate_last_binlog(last_binlog, pitr_ts, context, logger)
+
+            end_index, end_term = xdb_show_binlog_index(last_binlog, context, logger)
+            logger.info("end_index:%s;end_term:%s" % (end_index, end_term))
+
+            init_mysqld_metadata(cluster_start_index, commit_index, context, end_term, node_role, logger, is_pxc_xstore,
+                                 pitr_endpoint)
+
+            p = subprocess.Popen([
+                os.path.join(context.engine_home, 'bin', 'mysqld'),
+                "--defaults-file=" + context.mycnf_path,
+                "--user=mysql"
+            ], stdout=sys.stdout)
+
+            wait_binlog_apply_ready(context.port_access(), end_index, logger)
+
+            p.kill()
+            p.wait()
+        else:
+            chown_data_dir(context, logger)
+            init_mysqld_metadata(commit_index, commit_index, context, 0, node_role, logger, is_pxc_xstore,
+                                 pitr_endpoint)
     sync_cluster_metadata(context, logger)
-
     context.mark_node_initialized()
 
 
@@ -128,6 +171,21 @@ def mkdir_needed(context):
     shutil.chown(context.volume_path(VOLUME_DATA, "log"), "mysql", "mysql")
     shutil.chown(context.volume_path(VOLUME_DATA, "tmp"), "mysql", "mysql")
     shutil.chown(context.volume_path(VOLUME_DATA, "run"), "mysql", "mysql")
+    shutil.chown(context.volume_path(VOLUME_DATA), "mysql","mysql")
+
+
+def download_keyring_file(keyringfile_path,keyring_path, filestream_client, logger):
+    if len(keyring_path) != 0:
+        keyring_file_path = os.path.dirname(keyringfile_path)
+        logger.info("keyring_file_path:%s",keyring_file_path)
+        if not os.path.exists(keyring_file_path):
+            os.makedirs(keyring_file_path)
+        shutil.chown(keyring_file_path,"mysql","mysql")
+        keyring_path_local = os.path.join(keyring_file_path, "keyring")
+        filestream_client.download_to_file(remote=keyring_path, local=keyring_path_local, logger=logger)
+        logger.info("backup keyring downloaded!")
+        return keyring_path_local
+    return ""
 
 
 def download_backup_file(backup_file_path, backup_file_name, filestream_client, logger):
@@ -154,6 +212,7 @@ def download_pitr_binloglist(context, pitrEndpoint, xstore, logger):
     binlogListUrl = "/".join([pitrEndpoint, "binlogs"]) + ("?xstore=%s" % xstore)
     response = requests.get(binlogListUrl)
     mysql_binlog_list = []
+    pitr_ts = ""
     if response.status_code == 200:
         logger.info("binlogs http response %s" % response.content)
         binlogs = json.loads(response.content)
@@ -252,13 +311,13 @@ def initialize_local_mycnf(context: Context, logger):
     logger.info("local mycnf initialized!")
 
 
-def apply_backup_file(context, logger):
+def apply_backup_file(keyring_path_local, context, logger):
     # 应用全量备份集
     apply_backup_cmd = ""
     if context.is_galaxy80():
-        apply_backup_cmd = "%s --defaults-file=%s --prepare --target-dir=%s 2> %s/applybackup.log" \
+        apply_backup_cmd = "%s --defaults-file=%s --prepare --target-dir=%s --xtrabackup-plugin-dir=%s --keyring-file-data=%s 2> %s/applybackup.log" \
                            % (context.xtrabackup, context.mycnf_path, context.volume_path(VOLUME_DATA, 'data'),
-                              context.volume_path(VOLUME_DATA, "log"))
+                            context.xtrabackup_plugin, keyring_path_local,context.volume_path(VOLUME_DATA, "log"))
     elif context.is_xcluster57():
         apply_backup_cmd = "%s --defaults-file=%s --apply-log  %s 2> %s/applybackup.log" \
                            % (context.xtrabackup, context.mycnf_path, context.volume_path(VOLUME_DATA, 'data'),
@@ -272,23 +331,34 @@ def chown_data_dir(context, logger):
     check_run_process(['chown', '-R', 'mysql:mysql', context.volume_path(VOLUME_DATA, "data")], logger=logger)
 
 
-def init_mysqld_metadata(cluster_start_index, commit_index, context, end_term, pod_role, logger):
+def init_mysqld_metadata(cluster_start_index, commit_index, context, end_term, pod_role, logger, is_pxc_xstore,
+                         pitr_endpoint):
     if pod_role == NODE_ROLE_VOTER:
         cluster_start_index = commit_index
         if cluster_start_index is None:
             raise Exception("can NOT get logger commit index")
         logger.info("got logger commit index :%s" % cluster_start_index)
 
-    # mysqld 元数据初始化
-    init_metadata_cmd = [os.path.join(context.engine_home, 'bin', 'mysqld'),
-                         "--defaults-file=" + context.mycnf_path,
-                         "--cluster-current-term=" + str(end_term),
-                         "--cluster-info=" + context.xcluster_info_argument(local=True),
-                         "--cluster-force-change-meta=ON",
-                         "--cluster-force-single-mode=ON",
-                         "--loose-cluster-force-recover-index=" + str(cluster_start_index),
-                         "--cluster-start-index=" + str(cluster_start_index)
-                         ]
+    # mysqld 元数据初始化（标准版和企业版）
+    if is_pxc_xstore or len(pitr_endpoint) != 0:
+        init_metadata_cmd = [os.path.join(context.engine_home, 'bin', 'mysqld'),
+                             "--defaults-file=" + context.mycnf_path,
+                             "--cluster-current-term=" + str(end_term),
+                             "--cluster-info=" + context.xcluster_info_argument(local=True),
+                             "--cluster-force-change-meta=ON",
+                             "--cluster-force-single-mode=ON",
+                             "--loose-cluster-force-recover-index=" + str(cluster_start_index),
+                             "--cluster-start-index=" + str(cluster_start_index)
+                             ]
+    else:
+        init_metadata_cmd = [os.path.join(context.engine_home, 'bin', 'mysqld'),
+                             "--defaults-file=" + context.mycnf_path,
+                             "--cluster-info=" + context.xcluster_info_argument(local=True),
+                             "--cluster-force-change-meta=ON",
+                             "--cluster-force-single-mode=ON",
+                             "--loose-cluster-force-recover-index=" + str(cluster_start_index),
+                             "--cluster-start-index=" + str(cluster_start_index)
+                             ]
     logger.info("init_metadata_cmd" + str(init_metadata_cmd))
     check_run_process(init_metadata_cmd, logger=logger)
 
@@ -319,18 +389,24 @@ def get_xtrabackup_binlog_info_from_instance_local(context: Context):
 def wait_binlog_apply_ready(mysql_port, end_log_index, logger):
     timeout = 48 * 60 * 60
     deadline = time.time() + timeout
+    cnt = 0
     while time.time() < deadline:
         logger.info("wait applying binlog")
         try:
             time.sleep(10)
-            if check_binlog_apply_index_status(mysql_port, end_log_index, logger):
+            if check_binlog_apply_index_status(mysql_port, end_log_index, logger, False):
                 return
+            if cnt > 10:
+                # 兼容DN空binlog的bug
+                if check_binlog_apply_index_status(mysql_port, end_log_index, logger, True):
+                    return
+            cnt = cnt + 1
         except Exception as e:
             logger.info(e)
     raise TimeoutError("binlog apply timeout!")
 
 
-def check_binlog_apply_index_status(mysql_port, end_log_index, logger):
+def check_binlog_apply_index_status(mysql_port, end_log_index, logger, isEmptyBinlog):
     sql_list = "select * from information_schema.alisql_cluster_local"
     logger.info("Execute SQL: %s" % sql_list)
 
@@ -346,7 +422,9 @@ def check_binlog_apply_index_status(mysql_port, end_log_index, logger):
         logger.info("columns: %s" % columns)
         logger.info("last apply index: %s" % columns[-3])
         logger.info("end_log_index: %s" % end_log_index)
-        if int(columns[-3]) < int(end_log_index):
+        if isEmptyBinlog and int(columns[-3]) < int(end_log_index) - 1:
+            return False
+        if not isEmptyBinlog and int(columns[-3]) < int(end_log_index):
             return False
     return True
 
@@ -403,6 +481,35 @@ def show_last_and_first_binlog(context, logger):
     first_binlog = os.path.join(log_dir, first_binlog)
     logger.info("get last binlog:%s;first binlog:%s" % (last_binlog, first_binlog))
     return last_binlog, first_binlog
+
+
+def truncate_last_binlog(binlog_path, endTS, context, logger):
+    """
+    bb truncate mysql_bin.xxxxx --end-ts  --output cut.xxxxx
+    """
+    log_dir = context.volume_path(VOLUME_DATA, "log")
+    out_file = os.path.join(log_dir, "cut.out")
+
+    truncate_cmd = [context.bb_home, 'truncate', binlog_path,
+                    '--end-ts', str(endTS),
+                    '--output', out_file]
+    check_run_process(truncate_cmd, logger=logger)
+    shutil.chown(out_file, "mysql", "mysql")
+
+    os.rename(binlog_path, binlog_path + ".bak")
+    os.rename(out_file, binlog_path)
+
+
+def download_pitr_ts(pitrEndpoint, logger):
+    configUrl = "/".join([pitrEndpoint, "config"])
+    response = requests.get(configUrl)
+    if response.status_code == 200:
+        logger.info("config http response %s" % response.content)
+        config = json.loads(response.content)
+        pitr_ts = config['timestamp']
+    else:
+        raise Exception("failed to get config url = %s" % configUrl)
+    return pitr_ts
 
 
 def xdb_show_binlog_index(binlog_path, context, logger):
